pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-annotation-kit"
version = "0.1.0"
description = "Vision dataset labeling and conversion pipeline"
requires-python = ">=3.9"

dependencies = [
    "numpy",
    "matplotlib",
    "opencv-python",
    "pillow",
    "scipy",
    "torch",
    "ultralytics",
    "transformers @ git+https://github.com/huggingface/transformers.git",
]

[tool.setuptools.packages.find]
where = ["."]